+++ inbm/cloudadapter-agent/cloudadapter/cloud/adapter_factory.py
@@ -5,7 +5,6 @@ Copyright (C) 2017-2023 Intel Corporation
 SPDX-License-Identifier: Apache-2.0
 """
 
-
 from .adapters.azure_adapter import AzureAdapter
 from .adapters.telit_adapter import TelitAdapter
 from .adapters.generic_adapter import GenericAdapter
@@ -24,9 +23,8 @@ def load_adapter_config() -> Dict:
         with open(ADAPTER_CONFIG_PATH) as config_file:
             config_contents = config_file.read()
             return json.loads(config_contents)
-    except OSError as e:
+    except (OSError, ValueError) as e:
         raise BadConfigError(f"Could not load configuration: {e}")
-    return {}
 
 
 def get_adapter_config_filepaths() -> List[str]: