+++ inbm/cloudadapter-agent/cloudadapter/cloud/client/connections/_connection.py
@@ -5,7 +5,7 @@ Abstract base class used by all cloud connection objects.
 """
 
 import abc
-
+from typing import Optional
 
 class Connection(metaclass=abc.ABCMeta):  # pragma: no cover
 
@@ -14,6 +14,13 @@ class Connection(metaclass=abc.ABCMeta):  # pragma: no cover
         """A readonly property
 
         @return: (int) Current request ID
+        """
+        pass
+
+    def get_client_id(self) -> Optional[str]:
+        """A readonly property
+
+        @return: (int) Client Id
         """
         pass