+++ inbm/cloudadapter-agent/cloudadapter/cloud/cloud_builders.py
@@ -8,7 +8,7 @@ import logging
 
 from .client.connections.mqtt_connection import MQTTConnection
 from .client.messengers.one_way_messenger import OneWayMessenger
-from .client.handlers.recieve_respond_handler import RecieveRespondHandler
+from .client.handlers.recieve_respond_handler import ReceiveResponseHandler
 from .client.handlers.echo_handler import EchoHandler
 from .client.cloud_client import CloudClient
 from .client.utilities import ProxyConfig, TLSConfig, Formatter, MethodParser
@@ -111,23 +111,23 @@ def build_client_with_config(config: Dict[str, Any]) -> CloudClient:
         telemetry = build_messenger_with_config(telemetry)
     else:
         raise ClientBuildError(
-            "Missing 'attribute' information in the config to while setting up cloud connection.")
+            "Missing 'telemetry' information in the config to while setting up cloud connection.")
     if attribute:
         attribute = build_messenger_with_config(attribute)
     else:
         raise ClientBuildError(
-            "Missing MQTT config information while setting up cloud connection.")
+            "Missing 'attribute' MQTT config information while setting up cloud connection.")
     if event:
         event = build_messenger_with_config(event)
     else:
         raise ClientBuildError(
-            "Missing MQTT config information while setting up cloud connection.")
+            "Missing 'event' MQTT config information while setting up cloud connection.")
 
     # Build handler
     handler_config = config.get("method")
     if handler_config:
         parser_config = handler_config.get("parse")
-        handler = RecieveRespondHandler(
+        handler = ReceiveResponseHandler(
             topic_formatter=Formatter(
                 formatting=handler_config.get("pub"),
                 defaults=defaults),