+++ inbm/cloudadapter-agent/tests/unit/cloud/client/handlers/test_recieve_respond_handler.py
@@ -1,11 +1,11 @@
 """
-Unit tests for the RecieveRespondHandler
+Unit tests for the ReceiveResponseHandler
 
 
 """
 
 
-from cloudadapter.cloud.client.handlers.recieve_respond_handler import RecieveRespondHandler
+from cloudadapter.cloud.client.handlers.recieve_respond_handler import ReceiveResponseHandler
 from cloudadapter.cloud.client.connections._connection import Connection
 from cloudadapter.cloud.client.utilities import Formatter, MethodParser, MethodParsed
 
@@ -13,7 +13,7 @@ import unittest
 import mock
 
 
-class TestRecieveRespondHandler(unittest.TestCase):
+class TestReceiveResponseHandler(unittest.TestCase):
 
     def setUp(self):
         self.mock_connection = mock.create_autospec(Connection)
@@ -21,7 +21,7 @@ class TestRecieveRespondHandler(unittest.TestCase):
         self.mock_payload = mock.create_autospec(Formatter)
         self.mock_parser = mock.create_autospec(MethodParser)
 
-        self.recieve_respond_handler = RecieveRespondHandler(
+        self.receive_respond_handler = ReceiveResponseHandler(
             self.mock_topic,
             self.mock_payload,
             "subscribe_topic",
@@ -34,9 +34,9 @@ class TestRecieveRespondHandler(unittest.TestCase):
         self.mock_parser.parse.return_value = [MethodParsed(method="method")]
         mock_callback = mock.Mock()
 
-        self.recieve_respond_handler.bind("method", mock_callback)
+        self.receive_respond_handler.bind("method", mock_callback)
 
-        self.recieve_respond_handler._on_method("topic", "payload")
+        self.receive_respond_handler._on_method("topic", "payload")
         assert mock_callback.call_count == 1
 
     def test_on_method_exits_on_no_methods_succeeds(self):
@@ -44,7 +44,7 @@ class TestRecieveRespondHandler(unittest.TestCase):
         self.mock_payload.format.return_value = "payload"
         self.mock_parser.parse.return_value = []
 
-        self.recieve_respond_handler._on_method("topic", "payload")
+        self.receive_respond_handler._on_method("topic", "payload")
         assert self.mock_connection.publish.call_count == 0
 
     def test_on_method_exits_on_invalid_method_succeeds(self):
@@ -52,5 +52,5 @@ class TestRecieveRespondHandler(unittest.TestCase):
         self.mock_payload.format.return_value = "payload"
         self.mock_parser.parse.return_value = [MethodParsed(method="")]
 
-        self.recieve_respond_handler._on_method("topic", "payload")
+        self.receive_respond_handler._on_method("topic", "payload")
         assert self.mock_connection.publish.call_count == 0